requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
google-cloud-vision==3.4.4
python-dotenv==1.0.0
